+++ projects/angular-split/src/lib/component/split.component.ts
@@ -48,9 +48,9 @@ import { getInputPositiveNumber, getInputBoolean, isUserSizesValid, getPointFrom
                  class="as-split-gutter"
                  [style.flex-basis.px]="gutterSize"
                  [style.order]="index*2+1"
-                 (undetected.click)="clickGutter($event, index+1)"
-                 (undetected.mousedown)="startDragging($event, index*2+1, index+1)"
-                 (undetected.touchstart)="startDragging($event, index*2+1, index+1)">
+                 (as-split-undetected.click)="clickGutter($event, index+1)"
+                 (as-split-undetected.mousedown)="startDragging($event, index*2+1, index+1)"
+                 (as-split-undetected.touchstart)="startDragging($event, index*2+1, index+1)">
                 <div class="as-split-gutter-icon"></div>
             </div>
         </ng-template>`,