+++ src_app/app/app.module.ts
@@ -73,7 +73,7 @@ const routes = [
         CollapseModule.forRoot(),
         BsDropdownModule.forRoot(),
         SortableModule.forRoot(),
-        AngularSplitModule,
+        AngularSplitModule.forRoot(),
     ],
     providers: [{
         provide: LocationStrategy, useClass: HashLocationStrategy},