+++ src_app/app/component/home/home.route.component.ts
@@ -32,7 +32,7 @@ import { Component } from '@angular/core';
                     angular-split 
                     <a href="https://www.npmjs.com/package/angular-split"><img src="https://img.shields.io/npm/v/angular-split/latest.svg?style=flat-square" alt="npm latest version" height="18"></a>&nbsp;<a href="https://www.npmjs.com/package/angular-split"><img src="https://img.shields.io/npm/v/angular-split/next.svg?style=flat-square" alt="npm next version" height="18"></a>&nbsp;<a href="https://travis-ci.com/bertrandg/angular-split"><img src="https://img.shields.io/travis/com/bertrandg/angular-split/master.svg?style=flat-square&label=Cypress+E2E+tests" alt="Build Status" height="18"></a> 
                 </h1>
-                <h4>Lightweight Angular UI library to split views and allow dragging to resize areas using CSS flexbox layout.</h4>
+                <h4>Angular UI library to split views and allow dragging to resize areas using CSS flexbox layout.</h4>
                 <br>
                 <h5>Install npm module:</h5>
                 <pre [innerText]="code1"></pre>
@@ -52,7 +52,7 @@ export class HomeComponent {
   
 @NgModule({
   imports: [
-    AngularSplitModule,
+    AngularSplitModule.forRoot(),
     ...
   ],
   ...