+++ src_lazy_app/app/app.module.ts
@@ -21,9 +21,12 @@ const routes: Routes = [
   imports: [
     BrowserModule,
     RouterModule.forRoot(routes),
-    SharedModule,
+    SharedModule.forRoot(),
   ],
-  providers: [{provide: LocationStrategy, useClass: HashLocationStrategy}],
+  providers: [{
+    provide: LocationStrategy, 
+    useClass: HashLocationStrategy
+  }],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}